refactor(prune): Decouple interface and implementation
