+++ crates/prune/db/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "reth-prune"
+name = "reth-prune-db"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -15,11 +15,19 @@ workspace = true
 # reth
 reth-chainspec.workspace = true
 reth-exex-types.workspace = true
+reth-db-api.workspace = true
 reth-errors.workspace = true
 reth-provider.workspace = true
 reth-tokio-util.workspace = true
 reth-config.workspace = true
 reth-prune-types.workspace = true
+reth-primitives-traits.workspace = true
+reth-static-file-types.workspace = true
+reth-prune.workspace = true
+
+# ethereum
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
 
 # metrics
 reth-metrics.workspace = true
@@ -29,8 +37,17 @@ alloy-primitives.workspace = true
 # misc
 tracing.workspace = true
 thiserror.workspace = true
+itertools.workspace = true
+rayon.workspace = true
 tokio.workspace = true
+rustc-hash.workspace = true
 
 [dev-dependencies]
 # reth
+reth-db = { workspace = true, features = ["test-utils"] }
+reth-stages = { workspace = true, features = ["test-utils"] }
+reth-primitives-traits = { workspace = true, features = ["arbitrary"] }
 reth-testing-utils.workspace = true
+reth-tracing.workspace = true
+
+assert_matches.workspace = true
