+++ crates/prune/db/src/db_ext.rs
@@ -1,10 +1,10 @@
-use crate::PruneLimiter;
 use reth_db_api::{
     cursor::{DbCursorRO, DbCursorRW, RangeWalker},
     table::{Table, TableRow},
     transaction::DbTxMut,
     DatabaseError,
 };
+use reth_prune::PruneLimiter;
 use std::{fmt::Debug, ops::RangeBounds};
 use tracing::debug;
