+++ crates/prune/db/src/user/history.rs
