+++ crates/prune/prune/src/segments/mod.rs
@@ -1,23 +1,12 @@
-mod receipts;
 mod set;
-mod static_file;
-mod user;
 
 use crate::{PruneLimiter, PrunerError};
 use alloy_primitives::{BlockNumber, TxNumber};
 use reth_provider::{errors::provider::ProviderResult, BlockReader, PruneCheckpointWriter};
 use reth_prune_types::{PruneCheckpoint, PruneMode, PrunePurpose, PruneSegment, SegmentOutput};
 pub use set::SegmentSet;
-pub use static_file::{
-    Headers as StaticFileHeaders, Receipts as StaticFileReceipts,
-    Transactions as StaticFileTransactions,
-};
 use std::{fmt::Debug, ops::RangeInclusive};
 use tracing::error;
-pub use user::{
-    AccountHistory, MerkleChangeSets, Receipts as UserReceipts, ReceiptsByLogs, SenderRecovery,
-    StorageHistory, TransactionLookup,
-};
 
 /// A segment represents a pruning of some portion of the data.
 ///
@@ -53,14 +42,13 @@ pub trait Segment<Provider>: Debug + Send + Sync {
 }
 
 /// Segment pruning input, see [`Segment::prune`].
-#[derive(Debug)]
-#[cfg_attr(test, derive(Clone))]
+#[derive(Debug, Clone)]
 pub struct PruneInput {
-    pub(crate) previous_checkpoint: Option<PruneCheckpoint>,
+    pub previous_checkpoint: Option<PruneCheckpoint>,
     /// Target block up to which the pruning needs to be done, inclusive.
-    pub(crate) to_block: BlockNumber,
+    pub to_block: BlockNumber,
     /// Limits pruning of a segment.
-    pub(crate) limiter: PruneLimiter,
+    pub limiter: PruneLimiter,
 }
 
 impl PruneInput {
@@ -72,7 +60,7 @@ impl PruneInput {
     /// 2. If checkpoint doesn't exist, return 0.
     ///
     /// To get the range end: get last tx number for `to_block`.
-    pub(crate) fn get_next_tx_num_range<Provider: BlockReader>(
+    pub fn get_next_tx_num_range<Provider: BlockReader>(
         &self,
         provider: &Provider,
     ) -> ProviderResult<Option<RangeInclusive<TxNumber>>> {
@@ -118,7 +106,7 @@ impl PruneInput {
     /// 2. If checkpoint doesn't exist, use block 0.
     ///
     /// To get the range end: use block `to_block`.
-    pub(crate) fn get_next_block_range(&self) -> Option<RangeInclusive<BlockNumber>> {
+    pub fn get_next_block_range(&self) -> Option<RangeInclusive<BlockNumber>> {
         let from_block = self.get_start_next_block_range();
         let range = from_block..=self.to_block;
         if range.is_empty() {
@@ -132,7 +120,7 @@ impl PruneInput {
     ///
     /// 1. If checkpoint exists, use next block.
     /// 2. If checkpoint doesn't exist, use block 0.
-    pub(crate) fn get_start_next_block_range(&self) -> u64 {
+    pub fn get_start_next_block_range(&self) -> u64 {
         self.previous_checkpoint
             .and_then(|checkpoint| checkpoint.block_number)
             // Checkpoint exists, prune from the next block after the highest pruned one
