+++ crates/prune/prune/src/segments/set.rs
@@ -1,17 +1,4 @@
-use crate::segments::{
-    AccountHistory, MerkleChangeSets, ReceiptsByLogs, Segment, SenderRecovery, StorageHistory,
-    TransactionLookup, UserReceipts,
-};
-use alloy_eips::eip2718::Encodable2718;
-use reth_db_api::{table::Value, transaction::DbTxMut};
-use reth_primitives_traits::NodePrimitives;
-use reth_provider::{
-    providers::StaticFileProvider, BlockReader, ChainStateBlockReader, DBProvider,
-    PruneCheckpointReader, PruneCheckpointWriter, StaticFileProviderFactory,
-};
-use reth_prune_types::PruneModes;
-
-use super::{StaticFileHeaders, StaticFileReceipts, StaticFileTransactions};
+use crate::segments::Segment;
 
 /// Collection of [`Segment`]. Thread-safe, allocated on the heap.
 #[derive(Debug)]
@@ -42,60 +29,6 @@ impl<Provider> SegmentSet<Provider> {
     /// Consumes [`SegmentSet`] and returns a [Vec].
     pub fn into_vec(self) -> Vec<Box<dyn Segment<Provider>>> {
         self.inner
-    }
-}
-
-impl<Provider> SegmentSet<Provider>
-where
-    Provider: StaticFileProviderFactory<
-            Primitives: NodePrimitives<SignedTx: Value, Receipt: Value, BlockHeader: Value>,
-        > + DBProvider<Tx: DbTxMut>
-        + PruneCheckpointWriter
-        + PruneCheckpointReader
-        + BlockReader<Transaction: Encodable2718>
-        + ChainStateBlockReader,
-{
-    /// Creates a [`SegmentSet`] from an existing components, such as [`StaticFileProvider`] and
-    /// [`PruneModes`].
-    pub fn from_components(
-        static_file_provider: StaticFileProvider<Provider::Primitives>,
-        prune_modes: PruneModes,
-    ) -> Self {
-        let PruneModes {
-            sender_recovery,
-            transaction_lookup,
-            receipts,
-            account_history,
-            storage_history,
-            bodies_history: _,
-            merkle_changesets,
-            receipts_log_filter,
-        } = prune_modes;
-
-        Self::default()
-            // Static file headers
-            .segment(StaticFileHeaders::new(static_file_provider.clone()))
-            // Static file transactions
-            .segment(StaticFileTransactions::new(static_file_provider.clone()))
-            // Static file receipts
-            .segment(StaticFileReceipts::new(static_file_provider))
-            // Merkle changesets
-            .segment(MerkleChangeSets::new(merkle_changesets))
-            // Account history
-            .segment_opt(account_history.map(AccountHistory::new))
-            // Storage history
-            .segment_opt(storage_history.map(StorageHistory::new))
-            // User receipts
-            .segment_opt(receipts.map(UserReceipts::new))
-            // Receipts by logs
-            .segment_opt(
-                (!receipts_log_filter.is_empty())
-                    .then(|| ReceiptsByLogs::new(receipts_log_filter.clone())),
-            )
-            // Transaction lookup
-            .segment_opt(transaction_lookup.map(TransactionLookup::new))
-            // Sender recovery
-            .segment_opt(sender_recovery.map(SenderRecovery::new))
     }
 }
