+++ crates/stages/api/Cargo.toml
@@ -19,6 +19,7 @@ reth-network-p2p.workspace = true
 reth-tokio-util.workspace = true
 reth-consensus.workspace = true
 reth-prune.workspace = true
+reth-prune-db.workspace = true
 reth-errors.workspace = true
 reth-stages-types.workspace = true
 reth-static-file-types.workspace = true
